split up fuel core simulated node
